% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kissat-core.R
\name{kis_solve_literals}
\alias{kis_solve_literals}
\title{Solve using literals}
\usage{
kis_solve_literals(literals, max_solutions = 1)
}
\arguments{
\item{literals}{integer literals}

\item{max_solutions}{limit number of solutions}
}
\value{
Return NULL is problem is unsatisfiable.  Otherwise return a list
        of integer sequences of literals - each sequence is a solution.
}
\description{
Solve using literals
}
\examples{
kis_solve_literals(c(1, 2, 0, 2, -3, 0))
}
