% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kissat-core.R
\name{kis_version}
\alias{kis_version}
\title{Return a list of version information of the kisSAT library}
\usage{
kis_version()
}
\value{
Named list of version information. \code{kisSAT_API_VERSION} is the
        internal C API version, and \code{release} is the numbering of the 
        software release tarball.
}
\description{
Return a list of version information of the kisSAT library
}
\examples{
kis_version()
}
