% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kissatire-package.R
\docType{package}
\name{kissatire-package}
\alias{kissatire}
\alias{kissatire-package}
\title{kissatire: SAT Solver using 'Kissat'}
\description{
A solver for Boolean satisfiability problems ('SAT') using the 'Kissat' library. For this solver, 'SAT' problems can be formulated in two ways; using integer literals, or using logical expressions specified with the 'satire' package.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/coolbutuseless/kissatire}
  \item Report bugs at \url{https://github.com/coolbutuseless/kissatire/issues}
}

}
\author{
\strong{Maintainer}: Mike Cheng \email{mikefc@coolbutuseless.com} [copyright holder]

}
\keyword{internal}
